REF: refactor node creation logic into standalone function
